# api.yaml
openapi: 3.0.1
info:
  title: Exam Portal API
  version: 1.0.0
servers:
  - url: https://localhost:5001/api
...
paths:
  /exams:
    post:
      tags: [Exam]
      summary: Add a new exam
      operationId: addExam
      requestBody:
        description: Exam data to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamCreate"
        required: true
      responses:
        "201":
          description: Exam created
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Exam]
      summary: List all exams
      operationId: listExams
      parameters:
        - name: pageIndex
          in: query
          schema:
            type: integer
            format: int64
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int64
            default: 10
        - name: title
          in: query
          schema:
            type: string
            default: ""
        - name: startDateFrom
          in: query
          schema:
            type: string
            format: date
        - name: startDateTo
          in: query
          schema:
            type: string
            format: date
        - name: endDateFrom
          in: query
          schema:
            type: string
            format: date
        - name: endDateTo
          in: query
          schema:
            type: string
            format: date
        - name: createdBy
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  paging:
                    $ref: "#/components/schemas/paging"
                  exams:
                    type: array
                    items:
                      $ref: "#/components/schemas/Exam"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /exams/{id}:
    get:
      tags: [Exam]
      summary: Get an exam by ID
      operationId: getExamById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Exam details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
        "404":
          description: Exam not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Exam]
      summary: Update an existing exam
      operationId: updateExam
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        description: Exam data to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamUpdate"
        required: true
      responses:
        "200":
          description: Exam updated
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Exam not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Exam]
      summary: Delete an exam
      operationId: deleteExam
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Exam deleted
        "404":
          description: Exam not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /answers:
    post:
      tags: [Answer]
      summary: Add a new answer
      operationId: addAnswer
      requestBody:
        description: Answer data to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerCreate"
        required: true
      responses:
        "201":
          description: Answer created successfully
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Answer]
      summary: List all answers
      operationId: listAnswers
      parameters:
        - name: pageIndex
          in: query
          schema:
            type: integer
            format: int64
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int64
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  paging:
                    $ref: "#/components/schemas/paging"
                  answers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Answer"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /answers/{id}:
    get:
      tags: [Answer]
      summary: Get an answer by ID
      operationId: getAnswerById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Answer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "404":
          description: Answer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Answer]
      summary: Update an existing answer
      operationId: updateAnswer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        description: Answer data to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerUpdate"
        required: true
      responses:
        "200":
          description: Answer updated successfully

        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Answer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Answer]
      summary: Delete an answer
      operationId: deleteAnswer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Answer deleted
        "404":
          description: Answer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    paging:
      type: object
      properties:
        pageIndex:
          type: integer
          format: int64
          example: 1
        pageSize:
          type: integer
          format: int64
          example: 10
        totalPages:
          type: integer
          format: int64
          example: 100
        totalRecords:
          type: integer
          format: int64
          example: 100
    Exam:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
        description:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          format: duration
        totalMarks:
          type: integer
          nullable: true
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: integer
          nullable: true
        updatedBy:
          type: integer
          nullable: true
        deletedBy:
          type: integer
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        isDeleted:
          type: boolean
          nullable: true
    ExamCreate:
      type: object
      required: [title, durationMinutes, startDate, endDate]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          format: duration
        totalMarks:
          type: integer
          nullable: true
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdBy:
          type: integer
          nullable: true
    ExamUpdate:
      type: object
      required: [title, durationMinutes, startDate, endDate]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          format: duration
        totalMarks:
          type: integer
          nullable: true
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        updatedBy:
          type: integer
          nullable: true
...
    Answer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        questionId:
          type: integer
        selectedOptionId:
          type: integer
          nullable: true
    AnswerCreate:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer
        selectedOptionId:
          type: integer
          nullable: true
    AnswerUpdate:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer
        selectedOptionId:
          type: integer
          nullable: true
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string